avoid one liner
